% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprex.r
\name{do_nothing}
\alias{do_nothing}
\title{Reprex}
\usage{
do_nothing()
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Reprex of issues arising when recording graphical examples with \code{pdf(NULL)}
}
\examples{
heart_glyph <- "\u2665"
# Ideally this example would run in {pkgdown}
if (require("grid", quietly = TRUE) &&
    pkgdown.pdf.null.reprex:::device_supports_unicode()) {
  grid.newpage()
  gp = gpar(fontsize = 72, fontfamily = "sans")
  grid.text(heart_glyph, gp = gp)
}
if (require("grid") &&
    (pkgdown.pdf.null.reprex:::device_supports_unicode() ||
     pkgdown::in_pkgdown())) {
  grid.newpage()
  gp = gpar(fontsize = 72, fontfamily = "sans")
  grid.text(heart_glyph, gp = gp)
}
rook_glyph <- "\u2656"
font <- "Dejavu Sans"
if (require("grid") && 
    pkgdown.pdf.null.reprex:::has_font(font) &&
    (pkgdown.pdf.null.reprex:::device_supports_unicode() ||
     pkgdown::in_pkgdown())) {
  grid.newpage()
  gp = gpar(fontsize = 72, fontfamily = font)
  grid.text(rook_glyph, gp = gp)
}
}
